Move wtr fixtures
